+++ src/main/java/com/naocraftlab/foggypalegarden/domain/model/Color.java
@@ -1,0 +1,12 @@
+package com.naocraftlab.foggypalegarden.domain.model;
+
+import lombok.Builder;
+
+@Builder
+public record Color(
+        float red,
+        float green,
+        float blue,
+        float alpha
+) {
+}