+++ src/main/java/com/naocraftlab/foggypalegarden/domain/model/FogCharacteristics.java
@@ -1,0 +1,12 @@
+package com.naocraftlab.foggypalegarden.domain.model;
+
+import lombok.Builder;
+
+@Builder
+public record FogCharacteristics(
+        float startDistance,
+        float endDistance,
+        Color color,
+        float fogDensity
+) {
+}