+++ src/main/java/com/naocraftlab/foggypalegarden/domain/model/Weather.java
@@ -1,0 +1,7 @@
+package com.naocraftlab.foggypalegarden.domain.model;
+
+public enum Weather {
+    CLEAR,
+    RAIN,
+    THUNDER
+}