+++ src/main/java/com/naocraftlab/foggypalegarden/exception/FoggyPaleGardenEnvironmentException.java
@@ -1,0 +1,12 @@
+package com.naocraftlab.foggypalegarden.exception;
+
+public class FoggyPaleGardenEnvironmentException extends FoggyPaleGardenException {
+
+    public FoggyPaleGardenEnvironmentException(String message) {
+        super(message);
+    }
+
+    public FoggyPaleGardenEnvironmentException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}